Splitted control logic to controllers and services for unit testing
